Move database test helper methods back to database test case.
